+++ lib/cmds/store-config.ts
@@ -1,7 +1,7 @@
 import { Argv } from 'yargs';
 
 import { commands } from './store-config-cmds';
-import { KnownEnvironment, knownEnvironmentConfigurations } from '../helpers/openfga/fga.adapter';
+import { KnownEnvironment, knownEnvironmentConfigurations } from '../helpers/openfga/environment-config';
 
 exports.command = 'store-config [storeId]';
 exports.desc = 'Interact with store configuration';