+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.1.2
+
+### [0.1.2](https://github.com/rhamzeh/openfga-cli/compare/v0.1.1...v0.1.2) (2023-04-19)
+
+- fix: read latest authorization model when needed [a57ef6f](https://github.com/rhamzeh/openfga-cli/commit/a57ef6f2807f9a7d0d6ea0fc3309ddb4625bec09)
+- feat: allow dynamically creating a store when migrating [59d985d](https://github.com/rhamzeh/openfga-cli/commit/59d985d681e757212ec146ddb72ffdde9789c6af)
+
 ## v0.1.1
 
 ### [0.1.1](https://github.com/rhamzeh/openfga-cli/compare/v0.1.0...v0.1.1) (2023-04-10)